+++ src/main/java/org/morphix/convert/PropertyConversionEngine.java
@@ -14,9 +14,11 @@ package org.morphix.convert;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.morphix.convert.context.ConversionContext;
+import org.morphix.convert.function.SimpleConverter;
 import org.morphix.convert.strategy.PropertyArrayStrategy;
 import org.morphix.convert.strategy.PropertyBeanStrategy;
 import org.morphix.convert.strategy.PropertyCollectionStrategy;
@@ -45,12 +47,31 @@ public class PropertyConversionEngine implements ConversionEngine {
 	private final List<? extends PropertyConversionStrategy> strategies;
 
 	/**
+	 * A simple converter to convert property names. This can be used by strategies that need to convert property names as
+	 * part of their conversion process. The converter can be customized to apply specific naming conventions or
+	 * transformations to property names.
+	 */
+	private final SimpleConverter<String, String> propertyNameConverter;
+
+	/**
 	 * Constructs a new {@code PropertyConversionEngine} with the specified list of strategies.
 	 *
 	 * @param strategies the list of property conversion strategies to use
 	 */
 	public PropertyConversionEngine(final List<? extends PropertyConversionStrategy> strategies) {
-		this.strategies = strategies;
+		this(strategies, getDefaultPropertyNameConverter());
+	}
+
+	/**
+	 * Constructs a new {@code PropertyConversionEngine} with the specified list of strategies.
+	 *
+	 * @param strategies the list of property conversion strategies to use
+	 * @param propertyNameConverter a simple converter to convert property names, used by strategies that need to convert
+	 */
+	public PropertyConversionEngine(final List<? extends PropertyConversionStrategy> strategies,
+			final SimpleConverter<String, String> propertyNameConverter) {
+		this.strategies = Objects.requireNonNull(strategies, "Strategies list cannot be null");
+		this.propertyNameConverter = Objects.requireNonNull(propertyNameConverter, "Property name converter cannot be null");
 	}
 
 	/**
@@ -83,6 +104,27 @@ public class PropertyConversionEngine implements ConversionEngine {
 	}
 
 	/**
+	 * Returns the default list of property conversion strategies used by the default instance of
+	 * {@code PropertyConversionEngine}.
+	 *
+	 * @return the default list of property conversion strategies
+	 */
+	public static List<PropertyConversionStrategy> getDefaultStrategies() {
+		return InstanceHolder.DEFAULT_STRATEGIES;
+	}
+
+	/**
+	 * Returns the default property name converter used by the default instance of {@code PropertyConversionEngine}. The
+	 * default converter simply returns the input key as is, without applying any transformations. This can be used when no
+	 * custom property name converter is provided by the conversion engine.
+	 *
+	 * @return the default property name converter
+	 */
+	public static SimpleConverter<String, String> getDefaultPropertyNameConverter() {
+		return InstanceHolder.DEFAULT_PROPERTY_NAME_CONVERTER;
+	}
+
+	/**
 	 * Resolves the appropriate property conversion strategy for the given type by iterating through the list of strategies
 	 * and checking if each one supports the type. If a supporting strategy is found, it is returned. If no strategy
 	 * supports the type, an {@link IllegalStateException} is thrown.
@@ -110,6 +152,17 @@ public class PropertyConversionEngine implements ConversionEngine {
 	}
 
 	/**
+	 * Returns the simple converter used for converting property names. This converter can be used by strategies that need
+	 * to convert property names as part of their conversion process. The converter can be customized to apply specific
+	 * naming conventions or transformations to property names.
+	 *
+	 * @return the simple converter used for converting property names
+	 */
+	public SimpleConverter<String, String> getPropertyNameConverter() {
+		return propertyNameConverter;
+	}
+
+	/**
 	 * The default instance is lazily initialized when the getDefault() method is called for the first time. This approach
 	 * ensures thread safety and avoids unnecessary initialization if the default instance is never used.
 	 *
@@ -118,16 +171,30 @@ public class PropertyConversionEngine implements ConversionEngine {
 	private static class InstanceHolder {
 
 		/**
-		 * The default instance of {@code PropertyConversionEngine} initialized with a predefined set of strategies. The
-		 * strategies are ordered to ensure that more specific strategies (like leaf and optional) are applied before more
-		 * general ones (like collections and maps).
+		 * A default property name converter that simply returns the input key as is. This can be used when no custom property
+		 * name converter is provided by the conversion engine.
 		 */
-		private static final PropertyConversionEngine DEFAULT = new PropertyConversionEngine(List.of(
+		private static final SimpleConverter<String, String> DEFAULT_PROPERTY_NAME_CONVERTER = k -> k;
+
+		/**
+		 * The default list of property conversion strategies used by the default instance of {@code PropertyConversionEngine}.
+		 * The strategies are ordered to ensure that more specific strategies (like leaf and optional) are applied before more
+		 * general ones (like collections and maps). This order is important to ensure that the most appropriate strategy is
+		 * applied for each type of property value.
+		 */
+		private static final List<PropertyConversionStrategy> DEFAULT_STRATEGIES = List.of(
 				new PropertyLeafStrategy(),
 				new PropertyOptionalStrategy(),
 				new PropertyMapStrategy(),
 				new PropertyCollectionStrategy(),
 				new PropertyArrayStrategy(),
-				new PropertyBeanStrategy()));
+				new PropertyBeanStrategy());
+
+		/**
+		 * The default instance of {@code PropertyConversionEngine} initialized with a predefined set of strategies. The
+		 * strategies are ordered to ensure that more specific strategies (like leaf and optional) are applied before more
+		 * general ones (like collections and maps).
+		 */
+		private static final PropertyConversionEngine DEFAULT = new PropertyConversionEngine(DEFAULT_STRATEGIES);
 	}
 }